Pulling out log module
