+++ analysisengine/analysis-engine.go
@@ -5,8 +5,8 @@
 package analysisengine
 
 import (
-  "github.com/letsencrypt/boulder"
-  "github.com/streadway/amqp"
+	"github.com/letsencrypt/boulder/log"
+	"github.com/streadway/amqp"
 )
 
 // This file analyzes messages obtained from the Message Broker to determine
@@ -14,20 +14,20 @@ import (
 
 // Interface all Analysis Engines share
 type AnalysisEngine interface {
-  ProcessMessage(amqp.Delivery)
+	ProcessMessage(amqp.Delivery)
 }
 
 // An Analysis Engine that just logs to the JSON Logger.
 type LoggingAnalysisEngine struct {
-    jsonLogger *boulder.JsonLogger
+	jsonLogger *log.JsonLogger
 }
 
 func (eng *LoggingAnalysisEngine) ProcessMessage(delivery amqp.Delivery) {
-    // Send the entire message contents to the syslog server for debugging.
-    eng.jsonLogger.Debug("Message contents", delivery)
+	// Send the entire message contents to the syslog server for debugging.
+	eng.jsonLogger.Debug("Message contents", delivery)
 }
 
 // Construct a new Analysis Engine.
-func NewLoggingAnalysisEngine(logger *boulder.JsonLogger) AnalysisEngine {
-  return &LoggingAnalysisEngine{jsonLogger: logger}
+func NewLoggingAnalysisEngine(logger *log.JsonLogger) AnalysisEngine {
+	return &LoggingAnalysisEngine{jsonLogger: logger}
 }
